Gestion des stocks / TP2 début

+++ Taaroa_eats_flammekueche/TP1/TaaroaFFBaD/app/src/main/java/fr/utbm/lo52/taaroaffbad/Beans/Vente.java
@@ -13,13 +13,13 @@ public class Vente implements Serializable{
     private String marque;
     private String reference;
     private int acheteurId;
-    private int prix;
+    private double prix;
     private Boolean paye;
     private int quantite;
     private Date dateAchat;
     private Date datePaye;
 
-    public Vente(int VenteId, String marque, String reference, int fabricantId, int acheteurId, int prix, Boolean paye, int quantite, Date dateAchat, Date datePaye) {
+    public Vente(int VenteId, String marque, String reference, int fabricantId, int acheteurId, double prix, Boolean paye, int quantite, Date dateAchat, Date datePaye) {
         this.venteId = VenteId;
         this.fabricantId = fabricantId;
         this.marque = marque;
@@ -81,7 +81,7 @@ public class Vente implements Serializable{
         this.acheteurId = acheteurId;
     }
 
-    public int getPrix() {
+    public double getPrix() {
         return prix;
     }
